backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
firebase-admin==6.2.0
google-cloud-aiplatform==1.36.0
google-cloud-firestore==2.13.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
httpx==0.25.0
google-generativeai==0.3.1
